requirements.txt
alembic
asyncpg
fastapi
pillow
psycopg2
pydantic-settings
requests
sqlalchemy
sqlmodel
uvicorn